Add convenience run method
There are times where you are not interested in the result of a
future, but still you don't want to wrap the function and want to
execute it in a future chain.

With this method you can do that, where the supplied function shall be
executed in the callers thread and return void.

The same can be done by calling CompletableFuture.runAsync but with
the deference that this custom method will run in the callers thread.
If you need to execute something in the common fork join pool, or on
an executors pool, the CF library already provides mechanism for that.