Add obtain method to the Promise API
The obtain method shall ignore the previous completion stage return
value and just return the passed value it has or the passed function
return value. This is a syntax sugar method to avoid things like
`aVoid` or `dummy` parameters that are constantly passed around.

Also up the java build version to 11.